\docType{data}
\name{faces_dim}
\alias{faces_dim}
\title{vector telling how to unpack a vector from faces_mat into a 2-d array that can be plotted with \code{\link{image}}.}
\description{
vector telling how to unpack a vector from faces_mat into a
2-d array that can be plotted with \code{\link{image}}.
}
\details{
A vector containing the number of pixels per image, and
number of images
}
\references{
http://pics.psych.stir.ac.uk/2D_face_sets.htm
}
\keyword{data}

