% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/example_face.r
\docType{data}
\name{faces_mat}
\alias{faces_mat}
\title{Face matrix for example bootstrap PCA}
\description{
Face matrix for example bootstrap PCA
}
\details{
A matrix used from the analysis
}
\references{
http://pics.psych.stir.ac.uk/2D_face_sets.htm
}
\keyword{data}

