% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/example_face.r
\name{ex_faces}
\alias{ex_faces}
\title{Run eigenface bootstrap standard error example}
\usage{
ex_faces(K = 3, showFaceEx = FALSE, verbose = TRUE)
}
\arguments{
\item{K}{number of components to compute standard errors for}

\item{showFaceEx}{Whether examples faces from this dataset be plotted as illustrations of the data}
}
\value{
plots of the first K element-wise standard errors for the eigenfaces
}
\description{
Run eigenface bootstrap standard error example
}
\examples{
ex_faces()
}

